ADD: Inscriptions decorator
