+++ app/controllers/inscriptions_controller.rb
@@ -10,6 +10,7 @@ class InscriptionsController < ApplicationController
   # GET /inscriptions/1
   # GET /inscriptions/1.json
   def show
+    @inscription = @inscription.decorate
   end
 
   # GET /inscriptions/new
@@ -18,6 +19,7 @@ class InscriptionsController < ApplicationController
     @inscription = Inscription.new
     @inscription.event = @event
     @inscription.user = User.find(current_user.id)
+    @inscription = @inscription.decorate
   end
 
   # GET /inscriptions/1/edit
