+++ app/views/event_items/show.html.erb
@@ -6,7 +6,7 @@
     <h2 class="box-title">
       <i class="fa fa-fw fa-list"></i>
       <strong><%= @event_item.name %></strong>
-       pertence a
+      pertence a
       <%= link_to @event_item.event.name, @event_item.event, class: 'label label-default flat ' %>
     </h2>
   </div>
@@ -27,7 +27,9 @@
       </tr>
       <tr>
         <td width="20%"><strong><%= EventItem.human_attribute_name(:description) %>:</strong></td>
-        <td width="80%" colspan="3"> <pre><%= @event_item.description %></pre></td>
+        <td width="80%" colspan="3">
+          <pre><%= @event_item.description %></pre>
+        </td>
       </tr>
     </table>
   </div>
@@ -41,62 +43,62 @@
   <div class="box-body">
     <table class="table table-striped">
       <thead>
-      <td>
-        <th>
-          Nome
-        </th>
-        <th>
-          Função
-        </th>
-      </td>
+      <tr>
+      <th>
+        Nome
+      </th>
+      <th>
+        Função
+      </th>
+      </tr>
       </thead>
       <% @event_item.responsibles.each do |responsible| %>
         <tbody>
-        <td>
-          <tr>
+        <tr>
+          <td>
             <%= responsible.name %>
-          </tr>
-          <tr>
+          </td>
+          <td>
             <%= responsible.responsibility_type.name %>
-          </tr>
-        </td>
+          </td>
+        </tr>
         </tbody>
       <% end %>
     </table>
   </div>
-
-  <div class="box box-primary">
-  <div class="box-header with-border">
-    <h2 class="box-title">
-      <i class="fa fa-fw fa-list"></i> Inscritos
-    </h2>
-  </div>
-  <div class="box-body">
-    <table class="table table-striped">
-      <thead>
-      <td>
-        <th>
-          Nome
-        </th>
-        <th>
-          Email
-        </th>
-      </td>
-      </thead>
-      
-      <% @event_item.inscription_items.each do |inscription_activity| %>
-        <tbody>
-        <td>
-          <tr>
-            <%= inscription_activity.inscription.user.name %>
-          </tr>
-          <tr>
-            <%= inscription_activity.inscription.user.email %>
-          </tr>
-        </td>
-        </tbody>
-      <% end %>
-    </table>
-  </div>
-
 </div>
+  <div class="box box-primary">
+    <div class="box-header with-border">
+      <h2 class="box-title">
+        <i class="fa fa-fw fa-list"></i> Inscritos
+      </h2>
+    </div>
+    <div class="box-body">
+      <table class="table table-striped">
+        <thead>
+        <tr>
+          <th>
+            Nome
+          </th>
+          <th>
+            Email
+          </th>
+        </tr>
+        </thead>
+
+        <% @event_item.inscription_items.each do |inscription_activity| %>
+          <tbody>
+          <tr>
+            <td>
+              <%= inscription_activity.inscription.user.name %>
+            </td>
+            <td>
+              <%= inscription_activity.inscription.user.email %>
+            </td>
+          </tr>
+          </tbody>
+        <% end %>
+      </table>
+    </div>
+
+  </div>
