+++ test/decorators/event_decorator_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class EventDecoratorTest < Draper::TestCase
+end
