+++ test/decorators/event_item_decorator_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class EventItemDecoratorTest < Draper::TestCase
+end
