+++ test/decorators/inscription_decorator_test.rb
@@ -1,0 +1,4 @@
+require 'test_helper'
+
+class InscriptionDecoratorTest < Draper::TestCase
+end
